move reporting code out of elastic.py
